+++ src/RunCommand.php
@@ -2,14 +2,12 @@
 
 namespace Glamorous\Boiler;
 
-use SplFileInfo;
+use Glamorous\Boiler\Helpers\Template;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\Yaml\Exception\ParseException;
-use Symfony\Component\Yaml\Yaml;
 
 class RunCommand extends ConfigurationCommand
 {
@@ -29,8 +27,6 @@ class RunCommand extends ConfigurationCommand
 
     /**
      * Configure the command options.
-     *
-     * @throws BoilerException
      */
     protected function configure()
     {
@@ -70,149 +66,19 @@ class RunCommand extends ConfigurationCommand
     }
 
     /**
-     * Search for the template file and return it if found.
+     * Replace variables in given string.
      *
-     * @param string $name
+     * @param string $contents
      *
-     * @return null|SplFileInfo
+     * @return string
      */
-    protected function findTemplateFile(string $name): ?SplFileInfo
+    private function replaceVariables(string $contents): string
     {
-        $name .= '.yml';
-        $finder = new Finder();
-        foreach ($this->paths as $directory) {
-            $finder->files()->name('*.yml')->depth('== 0')->in($directory);
+        $variables = [
+            '{#PROJECT_NAME#}' => $this->variables['template_name'],
+        ];
 
-            foreach ($finder as $file) {
-                if ($file->getFilename() === $name) {
-                    return $file;
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Parse and load yml file from path.
-     *
-     * @param string $path
-     *
-     * @return mixed
-     *
-     * @throws BoilerException
-     */
-    protected function parseTemplateFile(string $path)
-    {
-        try {
-            return Yaml::parseFile($path);
-        } catch (ParseException $exception) {
-            throw new BoilerException(
-                'Template file located at `' . $path . '` could not be parsed',
-                $exception->getCode(),
-                $exception
-            );
-        }
-    }
-
-    /**
-     * Load included templates.
-     *
-     * @param array $template
-     *
-     * @return array
-     * @throws BoilerException
-     */
-    protected function getIncludedTemplates(array $template): array
-    {
-        if (!array_key_exists('include', $template)) {
-            return [];
-        }
-
-        if (!is_array($template['include'])) {
-            throw new BoilerException('Include function must be an array.');
-        }
-
-        return array_map(function ($fileName) {
-            $templateFile = $this->findTemplateFile($fileName);
-            if ($templateFile === null) {
-                throw new BoilerException('Included file `' . $fileName . '` does not exists');
-            }
-            $template = $this->parseTemplateFile($templateFile->getPathname());
-
-            if (! $template) {
-                throw new BoilerException('Included file `' . $fileName . '` cannot be parsed');
-            }
-
-            return $template;
-        }, $template['include']);
-    }
-
-    /**
-     * Validate the loaded template-file.
-     *
-     * @param array $template
-     *
-     * @throws BoilerException
-     */
-    protected function validateTemplate(array $template): void
-    {
-        if (!array_key_exists('name', $template) || empty($template['name'])) {
-            throw new BoilerException('There\'s no name defined in the template');
-        }
-
-        if (!array_key_exists('steps', $template) || !is_array($template['steps']) || empty($template['steps'])) {
-            throw new BoilerException('There are no steps defined in the template');
-        }
-
-        $this->validateSteps($template);
-    }
-
-    /**
-     * Merge the extra template (if there are) with the main template.
-     *
-     * @param array $template
-     * @param array $extraTemplates
-     *
-     * @return array
-     */
-    protected function mergeExtraTemplates(array $template, array $extraTemplates): array
-    {
-        foreach ($extraTemplates as $extraTemplate) {
-            $template = array_merge($template, $extraTemplate);
-        }
-
-        return $template;
-    }
-
-    /**
-     * Method to get a valid template from a given template name.
-     *
-     * @param string $templateFileName
-     *
-     * @return array
-     * @throws BoilerException
-     */
-    protected function getValidTemplateArray(string $templateFileName): array
-    {
-        $templateFile = $this->findTemplateFile($templateFileName);
-
-        if (is_null($templateFile)) {
-            throw new BoilerException('No template found with name ' . $templateFileName);
-        }
-
-        $template = $this->parseTemplateFile($templateFile->getPathname());
-
-        if (!is_array($template)) {
-            throw new BoilerException('Template could not be parsed as a yaml file');
-        }
-
-        $extraTemplates = $this->getIncludedTemplates($template);
-        $template = $this->mergeExtraTemplates($template, $extraTemplates);
-
-        $this->validateTemplate($template);
-
-        return $template;
+        return str_replace(array_keys($variables), array_values($variables), $contents);
     }
 
     /**
@@ -223,16 +89,11 @@ class RunCommand extends ConfigurationCommand
      */
     protected function executeScripts(OutputInterface $output, array $template): void
     {
-        $variables = [
-            '{#PROJECT_NAME#}' => $this->variables['template_name'],
-        ];
-
         foreach ($template['steps'] as $step) {
             $output->writeln('<info>Executing ' . $template[$step]['name'] . '</info>');
             $scripts = is_array($template[$step]['script']) ? $template[$step]['script'] : [$template[$step]['script']];
             foreach ($scripts as $script) {
-                $script = str_replace(array_keys($variables), array_values($variables), $script);
-                exec($script);
+                exec($this->replaceVariables($script));
             }
         }
     }
@@ -254,7 +115,7 @@ class RunCommand extends ConfigurationCommand
         }
 
         $templateFileName = $input->getArgument('template');
-        $template = $this->getValidTemplateArray($templateFileName);
+        $template = Template::getInstance()->searchTemplateByFilenameInGivenPaths($templateFileName, $this->paths);
 
         $templateName = $input->getOption('name') ? $input->getOption('name') : $template['name'];
         $directoryName = $input->getOption('dir') ?: $templateFileName;
@@ -269,31 +130,42 @@ class RunCommand extends ConfigurationCommand
         mkdir($directoryName);
         chdir($directoryName);
 
+        $this->prepareTemplateFolder($templateFileName);
+
         $this->executeScripts($output, $template);
     }
 
     /**
-     * Internal method to validate the steps.
+     * Prepare the project directory by copying files from the template directory if available.
      *
-     * @param array $template
-     *
-     * @throws BoilerException
+     * @param string $name
      */
-    private function validateSteps(array $template): void
+    protected function prepareTemplateFolder(string $name): void
     {
-        $steps = $template['steps'];
-        foreach ($steps as $step) {
-            if (!array_key_exists($step, $template)) {
-                throw new BoilerException('The step `' . $step . '` does not exists.');
+        $finder = new Finder();
+        foreach ($this->paths as $directory) {
+            $templateDir = $directory . '/' . $name;
+            if (! is_dir($templateDir)) {
+                continue;
             }
 
-            if (!array_key_exists('name', $template[$step])) {
-                throw new BoilerException('No `name` set for the step `' . $step . '`');
-            }
+            $finder->in($templateDir)
+                ->notName($name . '.yml')
+                ->sortByType()
+                ->ignoreDotFiles(false);
 
-            if (!array_key_exists('script', $template[$step])) {
-                throw new BoilerException('No `script` set for the step `' . $step . '`');
+            foreach ($finder as $file) {
+                if ($file->isDir()) {
+                    mkdir($file->getFilename(), 0777, true);
+                    continue;
+                }
+
+                if ($file->isFile()) {
+                    copy($file->getPathname(), $file->getRelativePathname());
+                    file_put_contents($file->getRelativePathname(), $this->replaceVariables($file->getContents()));
+                    continue;
+                }
             }
-        }
+        }//end foreach
     }
 }