+++ src/SetupPathCommand.php
@@ -2,8 +2,6 @@
 
 namespace Glamorous\Boiler;
 
-use Exception;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;